+++ shared/src/lib.rs
@@ -31,9 +31,10 @@ where
 
 mod base_image;
 mod download_ruby_version;
+pub mod github;
 mod inventory_help;
 pub mod maybe_err;
-pub mod github;
+pub mod s3;
 
 pub use base_image::{BaseImage, base_images, build_matrix};
 pub use download_ruby_version::RubyDownloadVersion;
